+++ deepchem/feat/complex_featurizers/__init__.py
@@ -1,0 +1,3 @@
+# flake8: noqa
+from deepchem.feat.complex_featurizers.atomic_coordinates import NeighborListComplexAtomicCoordinates
+from deepchem.feat.complex_featurizers.atomic_coordinates import ComplexNeighborListFragmentAtomicCoordinates
