+++ deepchem/feat/tests/test_rdkit_descriptors.py
@@ -4,7 +4,7 @@ Test basic molecular features.
 import numpy as np
 import unittest
 
-from deepchem.feat.rdkit_descriptors import RDKitDescriptors
+from deepchem.feat import RDKitDescriptors
 
 
 class TestRDKitDescriptors(unittest.TestCase):
